some cleanup and added minor functionalities - moving towards having a common set of matrix utils here

+++ py_factor_graph/calibration.py
@@ -98,7 +98,7 @@ class RangePairCalibration:
     @property
     def bias(self) -> float:
         """The bias of the range measurement."""
-        return np.mean(self.measured_dists) - np.mean(self.true_dists)
+        return float(np.mean(self.measured_dists) - np.mean(self.true_dists))
 
     @property
     def covariance(self):
